Add a CI job to Console Static Text
Adds a GitHub Actions job to run the build & tests on every push. The
project is built with all feature flag enabled.

This change-set also adds a variety of test cases for the tokenizer in
`words.rs`. To make testing easier, the `PartialEq` and `Debug` debug
traits were added to the `WordToken` enum.